W2-G3: plain floats at optimizer boundary; full schedule horizon (#283, #284)
Coerce compute_solar_gain and _deferred_energy_cost to plain Python float
so predicted_savings and schedule solar_gain are not numpy.float64 leaves.
Publish the full solve horizon in legacy schedule/dhw_schedule instead of
hard-coded 24 steps. Witnessed by features.py and D8/matrix.py at 757e164.

+++ custom_components/heatpump_optimizer/optimizer.py
@@ -5344,7 +5344,7 @@ class HeatPumpOptimizer:
                     - _capped(optimized_end.dhw_temperature, "dhw")
                 )
                 electrical += dhw_gap / cop_dhw - dhw_gap / cop
-        return electrical * refill_price
+        return float(electrical * refill_price)
 
     def _compute_baseline_power(
         self,
@@ -5877,7 +5877,9 @@ class HeatPumpOptimizer:
 
         # Add solar gain info
         if result.solar_gain_trajectory and i < len(result.solar_gain_trajectory):
-            action["solar_gain_kw"] = round(result.solar_gain_trajectory[i], 3)
+            action["solar_gain_kw"] = float(
+                round(result.solar_gain_trajectory[i], 3)
+            )
 
         # Add DHW info
         if result.dhw_power_schedule and i < len(result.dhw_power_schedule):
